+++ coeus-ops/src/reduction/mod.rs
@@ -13,8 +13,8 @@ pub use cumsum::{cumsum, suffix_sum};
 pub use linalg::{cross, dot};
 pub use mean::{mean, mean_axis};
 pub use stats::{
-    frobenius_norm, frobenius_norm_batched, norm, norm_p, norm_p_axis, std_dev, std_dev_axis, var,
-    var_axis,
+    frobenius_norm, frobenius_norm_batched, norm, norm_p, norm_p_axis, std_dev, std_dev_axis,
+    std_mean, std_mean_axis, var, var_axis, var_mean, var_mean_axis,
 };
 pub use sum::{amax, amin, max_axis, min_axis, prod, sum, sum_axis};
 pub use topk::{argmax, argmin, topk};